requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.1.7
dp==1.0
exceptiongroup==1.3.0
executor==23.2
fasteners==0.19
fonttools==4.57.0
frozenlist==1.5.0
geomet==0.2.1.post1
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
image==1.5.33
jiter==0.9.0
joblib==1.4.2
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.1
multidict==6.1.0
numpy==2.2.5
openai==1.71.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
property-manager==3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.17.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.2
undetected-chromedriver==3.5.3
urllib3==2.3.0
verboselogs==1.7
webdriver-manager==4.0.2
websockets==15.0.1
wsproto==1.2.0
yarl==1.18.3
